backend/requirements.txt
asgiref==3.8.1
Django==4.2.20
sqlparse==0.5.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
djangorestframework==3.16.0
django-cors-headers==4.7.0